move circuit-value.ts
