+++ src/index.ts
@@ -21,7 +21,7 @@ export type {
   FlexibleProvable,
   FlexibleProvablePure,
   InferProvable,
-} from './lib/circuit-value.js';
+} from './lib/provable-types/circuit-value.js';
 export {
   CircuitValue,
   prop,
@@ -31,7 +31,7 @@ export {
   provablePure,
   Struct,
   Unconstrained,
-} from './lib/circuit-value.js';
+} from './lib/provable-types/circuit-value.js';
 export { Provable } from './lib/provable.js';
 export { Circuit, Keypair, public_, circuitMain } from './lib/circuit.js';
 export { UInt32, UInt64, Int64, Sign, UInt8 } from './lib/int.js';
