+++ src/lib/foreign-ecdsa.ts
@@ -1,6 +1,6 @@
 import { provableFromClass } from '../bindings/lib/provable-snarky.js';
 import { CurveParams } from '../bindings/crypto/elliptic-curve.js';
-import { ProvablePureExtended } from './circuit-value.js';
+import { ProvablePureExtended } from './provable-types/circuit-value.js';
 import {
   FlexiblePoint,
   ForeignCurve,
