+++ src/lib/hash.ts
@@ -1,4 +1,8 @@
-import { HashInput, ProvableExtended, Struct } from './circuit-value.js';
+import {
+  HashInput,
+  ProvableExtended,
+  Struct,
+} from './provable-types/circuit-value.js';
 import { Snarky } from '../snarky.js';
 import { Field } from './core.js';
 import { createHashHelpers } from './hash-generic.js';
