+++ src/lib/provable-types/circuit-value.ts
@@ -1,24 +1,24 @@
 import 'reflect-metadata';
-import { Snarky } from '../snarky.js';
-import { Field, Bool, Scalar, Group } from './core.js';
+import { Snarky } from '../../snarky.js';
+import { Field, Bool, Scalar, Group } from '../core.js';
 import {
   provable,
   provablePure,
   provableTuple,
   HashInput,
   NonMethods,
-} from '../bindings/lib/provable-snarky.js';
+} from '../../bindings/lib/provable-snarky.js';
 import type {
   InferJson,
   InferProvable,
   InferredProvable,
   IsPure,
-} from '../bindings/lib/provable-snarky.js';
-import { Provable } from './provable.js';
-import { assert } from './errors.js';
-import { inCheckedComputation } from './provable-context.js';
-import { Proof } from './proof-system.js';
-import { ProvablePure } from './provable-types/provable-intf.js';
+} from '../../bindings/lib/provable-snarky.js';
+import { Provable } from '../provable.js';
+import { assert } from '../errors.js';
+import { inCheckedComputation } from '../provable-context.js';
+import { Proof } from '../proof-system.js';
+import { ProvablePure } from './provable-intf.js';
 
 // external API
 export {
